add support for agc-manager
- The agc-manager is part of the alpaka-group-container and allow to search for preinstalled software.
- replace ALPAKA_CI_BOOST_BRANCH with ALPAKA_BOOST_VERSION

+++ script/install_clang.sh
@@ -18,52 +18,55 @@ source ./script/set.sh
 : "${ALPAKA_CI_STDLIB?'ALPAKA_CI_STDLIB must be specified'}"
 : "${CXX?'CXX must be specified'}"
 
-# Install from LLVM repository (if available); otherwise install LLVM from official Ubuntu repositories
-ALPAKA_CI_UBUNTU_NAME=`lsb_release -c | awk '{print $2}'`
-wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
-
-# focal = 20.04; jammy = 22.04
-if { [ "${ALPAKA_CI_UBUNTU_NAME}" == "focal" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 9 ]; } || \
-   { [ "${ALPAKA_CI_UBUNTU_NAME}" == "jammy" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 13 ]; }
+if ! agc-manager -e clang@${ALPAKA_CI_CLANG_VER}
 then
-    sudo add-apt-repository "deb http://apt.llvm.org/${ALPAKA_CI_UBUNTU_NAME}/ llvm-toolchain-${ALPAKA_CI_UBUNTU_NAME}-$ALPAKA_CI_CLANG_VER main"
-fi
+    # Install from LLVM repository (if available); otherwise install LLVM from official Ubuntu repositories
+    ALPAKA_CI_UBUNTU_NAME=`lsb_release -c | awk '{print $2}'`
+    wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | sudo apt-key add -
 
-travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install clang-${ALPAKA_CI_CLANG_VER}
-
-if [ -n "${ALPAKA_CI_SANITIZERS}" ]
-then
-    # llvm-symbolizer is required for meaningful output. This is part of the llvm base package which we don't install by default.
-    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install llvm-${ALPAKA_CI_CLANG_VER}
-fi
-
-if [ "${ALPAKA_CI_STDLIB}" == "libc++" ]
-then
-    travis_retry sudo apt-get -y --quiet update
-    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++-${ALPAKA_CI_CLANG_VER}-dev
-    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++abi-${ALPAKA_CI_CLANG_VER}-dev
-    if [ "${ALPAKA_CI_CLANG_VER}" -ge 12 ]
+    # focal = 20.04; jammy = 22.04
+    if { [ "${ALPAKA_CI_UBUNTU_NAME}" == "focal" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 9 ]; } || \
+    { [ "${ALPAKA_CI_UBUNTU_NAME}" == "jammy" ] && [ "${ALPAKA_CI_CLANG_VER}" -ge 13 ]; }
     then
-        # Starting from LLVM 12 libunwind is required when using libc++. For some reason this isn't installed by default
-        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libunwind-${ALPAKA_CI_CLANG_VER}-dev
+        sudo add-apt-repository "deb http://apt.llvm.org/${ALPAKA_CI_UBUNTU_NAME}/ llvm-toolchain-${ALPAKA_CI_UBUNTU_NAME}-$ALPAKA_CI_CLANG_VER main"
     fi
-fi
 
-if [ "${alpaka_ACC_CPU_B_OMP2_T_SEQ_ENABLE}" = "ON" ] || [ "${alpaka_ACC_CPU_B_SEQ_T_OMP2_ENABLE}" = "ON" ] || [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
-then
-    LIBOMP_PACKAGE=libomp-${ALPAKA_CI_CLANG_VER}-dev
-    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install "${LIBOMP_PACKAGE}"
-    if [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
+    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install clang-${ALPAKA_CI_CLANG_VER}
+
+    if [ -n "${ALPAKA_CI_SANITIZERS}" ]
     then
-        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install \
-            clang-tools-${ALPAKA_CI_CLANG_VER} llvm-${ALPAKA_CI_CLANG_VER}
+        # llvm-symbolizer is required for meaningful output. This is part of the llvm base package which we don't install by default.
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install llvm-${ALPAKA_CI_CLANG_VER}
     fi
-fi
 
-sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/cc cc /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
-sudo update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+    if [ "${ALPAKA_CI_STDLIB}" == "libc++" ]
+    then
+        travis_retry sudo apt-get -y --quiet update
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++-${ALPAKA_CI_CLANG_VER}-dev
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libc++abi-${ALPAKA_CI_CLANG_VER}-dev
+        if [ "${ALPAKA_CI_CLANG_VER}" -ge 12 ]
+        then
+            # Starting from LLVM 12 libunwind is required when using libc++. For some reason this isn't installed by default
+            travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install libunwind-${ALPAKA_CI_CLANG_VER}-dev
+        fi
+    fi
+
+    if [ "${alpaka_ACC_CPU_B_OMP2_T_SEQ_ENABLE}" = "ON" ] || [ "${alpaka_ACC_CPU_B_SEQ_T_OMP2_ENABLE}" = "ON" ] || [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
+    then
+        LIBOMP_PACKAGE=libomp-${ALPAKA_CI_CLANG_VER}-dev
+        travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install "${LIBOMP_PACKAGE}"
+        if [ "${alpaka_ACC_ANY_BT_OMP5_ENABLE}" = "ON" ]
+        then
+            travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install \
+                clang-tools-${ALPAKA_CI_CLANG_VER} llvm-${ALPAKA_CI_CLANG_VER}
+        fi
+    fi
+
+    sudo update-alternatives --install /usr/bin/clang clang /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/clang++ clang++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/cc cc /usr/bin/clang-"${ALPAKA_CI_CLANG_VER}" 50
+    sudo update-alternatives --install /usr/bin/c++ c++ /usr/bin/clang++-"${ALPAKA_CI_CLANG_VER}" 50
+fi
 
 which "${CXX}"
 ${CXX} -v
