+++ script/install_gcc.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #
-# Copyright 2017-2019 Benjamin Worpitz
+# Copyright 2022 Benjamin Worpitz, Simeon Ehrig
 #
 # This file is part of alpaka.
 #
@@ -18,10 +18,13 @@ source ./script/set.sh
 : "${ALPAKA_CI_SANITIZERS?'ALPAKA_CI_SANITIZERS must be specified'}"
 : "${CXX?'CXX must be specified'}"
 
-travis_retry sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
-travis_retry sudo apt-get -y --quiet update
+if ! agc-manager -e gcc@${ALPAKA_CI_GCC_VER}
+then
+    travis_retry sudo add-apt-repository -y ppa:ubuntu-toolchain-r/test
+    travis_retry sudo apt-get -y --quiet update
+    travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install g++-"${ALPAKA_CI_GCC_VER}"
+fi
 
-travis_retry sudo apt-get -y --quiet --allow-unauthenticated --no-install-recommends install g++-"${ALPAKA_CI_GCC_VER}"
 sudo update-alternatives --install /usr/bin/gcc gcc /usr/bin/gcc-"${ALPAKA_CI_GCC_VER}" 50
 sudo update-alternatives --install /usr/bin/g++ g++ /usr/bin/g++-"${ALPAKA_CI_GCC_VER}" 50
 if [[ "${ALPAKA_CI_SANITIZERS}" == *"TSan"* ]]
