+++ script/install_oneapi.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 #
-# Copyright 2020 Axel Huebl
+# Copyright 2022 Axel Huebl, Simeon Ehrig
 #
 # This file is part of alpaka.
 #
@@ -16,28 +16,32 @@ source ./script/set.sh
 
 : "${CXX?'CXX must be specified'}"
 
-# Ref.: https://github.com/rscohn2/oneapi-ci
-# intel-basekit intel-hpckit are too large in size
 
-travis_retry sudo apt-get -qqq update
-travis_retry sudo apt-get install -y wget build-essential pkg-config cmake ca-certificates gnupg
-travis_retry sudo wget https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
+if ! agc-manager -e oneapi
+then
+    # Ref.: https://github.com/rscohn2/oneapi-ci
+    # intel-basekit intel-hpckit are too large in size
 
-sudo apt-key add GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
-echo "deb https://apt.repos.intel.com/oneapi all main" | sudo tee /etc/apt/sources.list.d/oneAPI.list
+    travis_retry sudo apt-get -qqq update
+    travis_retry sudo apt-get install -y wget build-essential pkg-config cmake ca-certificates gnupg
+    travis_retry sudo wget https://apt.repos.intel.com/intel-gpg-keys/GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
 
-travis_retry sudo apt-get update
+    sudo apt-key add GPG-PUB-KEY-INTEL-SW-PRODUCTS-2023.PUB
+    echo "deb https://apt.repos.intel.com/oneapi all main" | sudo tee /etc/apt/sources.list.d/oneAPI.list
 
-#  See a list of oneAPI packages available for install
-echo "################################"
-sudo -E apt-cache pkgnames intel
-echo "################################"
+    travis_retry sudo apt-get update
 
-travis_retry sudo apt-get install -y intel-oneapi-compiler-dpcpp-cpp-and-cpp-classic intel-oneapi-mkl-devel intel-oneapi-openmp intel-oneapi-tbb-devel
+    #  See a list of oneAPI packages available for install
+    echo "################################"
+    sudo -E apt-cache pkgnames intel
+    echo "################################"
 
-set +eu
-source /opt/intel/oneapi/setvars.sh
-set -eu
+    travis_retry sudo apt-get install -y intel-oneapi-compiler-dpcpp-cpp-and-cpp-classic intel-oneapi-mkl-devel intel-oneapi-openmp intel-oneapi-tbb-devel
+
+    set +eu
+    source /opt/intel/oneapi/setvars.sh
+    set -eu
+fi
 
 which "${CXX}"
 ${CXX} -v
